requirements.txt
Fabric
Flask-SQLAlchemy
Flask-WTF
WTForms
coverage
requests
flask_migrate
flask_script
# ecdsa
# paramiko
# pycrypto
